encrypted_health_prediction/requirements.txt
concrete-ml==1.4.0
gradio
uvicorn>=0.21.0
fastapi>=0.93.0
